moved the component dir to root level
