+++ app/ssiForm/patient_data.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import InputBox from '../components/InputBox';
-import DropdownBox from '../components/DropdownBox';
-import DateTimePickerBox from '../components/DateTimePickerBox';
+import InputBox from '../../components/InputBox';
+import DropdownBox from '../../components/DropdownBox';
+import DateTimePickerBox from '../../components/DateTimePickerBox';
 import { FormData } from './page';
 
 interface PatientDataProps {
